further work

+++ src/gtfsparser/gtfs/shape.h
@@ -17,12 +17,12 @@ namespace gtfsparser {
 namespace gtfs {
 
 struct ShapePoint {
-  ShapePoint(double lat, double ln, double dist, uint16_t seq)
+  ShapePoint(double lat, double ln, double dist, uint32_t seq)
   : lat(lat), lng(ln), travelDist(dist), seq(seq) {}
   ShapePoint()
   : lat(0), lng(0), travelDist(-1), seq(0) {}
   double lat, lng, travelDist;
-  uint16_t seq;
+  uint32_t seq;
 };
 
 struct ShapePointCompare {
